Move Raveny bits where they belong
