+++ gore/tests/mock_raven.py
